refactor to extract, add line to yaml
